% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{addCorrelationKable}
\alias{addCorrelationKable}
\title{Display corelation result generate with getCorrelationCoeficient()}
\usage{
addCorrelationKable(data_frame)
}
\arguments{
\item{data_frame}{a data.frame, corelation result generate with getCorrelationCoeficient()}
}
\value{
an html data table to display
}
\description{
Display corelation result generate with getCorrelationCoeficient() as kableExtra object
}
\examples{
data(mtcars)
cor <- getCorrelationCoeficient(data = mtcars, var1 = 'mpg', var2 = colnames(mtcars)[sapply(mtcars, is.numeric)])
cor <- do.call('rbind', cor)
addCorrelationKable(cor)
}
