% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{addDataDescriptionTable}
\alias{addDataDescriptionTable}
\title{Add table of description of i2b2 data into a rdocx element}
\usage{
addDataDescriptionTable(doc, table, title)
}
\arguments{
\item{doc}{a rdocx element}

\item{table}{a data.frame, containing description of i2b2data}

\item{title}{a character vector of length one, the title of the table to display}
}
\value{
a rdocx element with the table at the end of the document
}
\description{
Add table of description of i2b2 data into a rdocx element
}
