% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_function.R
\name{addKable_html}
\alias{addKable_html}
\title{A convenient method to add kable into a RMarkdown document}
\usage{
addKable_html(data_frame, all_before = FALSE, var.width = 50, caption = NULL)
}
\arguments{
\item{data_frame}{a data.frame, containing the data to add}

\item{all_before}{a boolean vector of length one, indicatig if all column must be before strata column. Default to FALSE}

\item{var.width}{a numeric vector of length one, containing the witdh of the variable column}

\item{caption}{a character vector of length one, containing the caption of the table}
}
\value{
a knitr::kable
}
\description{
A convenient method to add kable into a RMarkdown document
}
