% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addLabelToVariable}
\alias{addLabelToVariable}
\title{Set label to variable in a data.frame}
\usage{
addLabelToVariable(.data, ..., .labels = NA)
}
\arguments{
\item{.data}{a data.frame}
}
\value{
a data.frame with variable label set
}
\description{
Set label to variable in a data.frame

Set label to variable in a data.frame
}
