% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{addNewLine}
\alias{addNewLine}
\title{Add new line to rdocx element}
\usage{
addNewLine(doc)
}
\arguments{
\item{doc}{a rdocx element}
}
\value{
a rdocx element with a new line at the end
}
\description{
Add a new line to rdocx element
}
\examples{
data(mtcars)
mtcars$vs <- as.factor(mtcars$vs)
plot <- ggplot2::ggplot(mtcars, ggplot2::aes(x = vs, y=mpg)) +
  ggplot2::geom_boxplot()

doc <- read_docx()
doc <- addTextElement(doc, "Titre 1 : description of the dataset", style = "heading 1")
doc <- addNewLine(doc)
doc <- addNewLine(doc)
doc <- addPlot(doc, plot, "mpg function of vs")
print(doc, target = "output.docx")
}
