% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{addPlot}
\alias{addPlot}
\title{Add plot element into a rdocx element}
\usage{
addPlot(doc, plot, title, width = 7, height = 7)
}
\arguments{
\item{doc}{a rdocx element}

\item{plot}{a plot element}

\item{title}{a character vector of length one, containing the totre of the plot to display}

\item{width}{a double vector of length one, the width in inch of the plot}

\item{height}{a double vector of length one, the height in inch of the plot}
}
\value{
a rdocx element with a plot at the end
}
\description{
Add plot element into a rdocx element
}
\examples{
data(mtcars)
mtcars$vs <- as.factor(mtcars$vs)
plot <- ggplot2::ggplot(mtcars, ggplot2::aes(x = vs, y=mpg)) +
  ggplot2::geom_boxplot()
doc <- read_docx()
doc <- addPlot(doc, plot, "mpg function of vs")
print(doc, target = "output.docx")
}
