% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_function.R
\name{addRegressionKable}
\alias{addRegressionKable}
\title{A convenient method to add regression kable into a RMarkdown document}
\usage{
addRegressionKable(data_frame)
}
\arguments{
\item{data_frame}{a data.frame, containing the data to add}
}
\value{
a knitr::kable
}
\description{
A convenient method to add regression kable into a RMarkdown document
}
