% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{addRegressionTable}
\alias{addRegressionTable}
\title{Add table into a rdocx element}
\usage{
addRegressionTable(doc, table, title, modalitySize = 1, sampleSize = 1,
  parameterSize = 1, confintSize = 2)
}
\arguments{
\item{doc}{a rdocx element}

\item{table}{a data.frame generated by statsBordeaux package}

\item{title}{a character vector of length one, the title of the table to display}

\item{modalitySize}{double, size in inch of the modality column}
}
\value{
a rdocx element with the table at the end of the document
}
\description{
Add table generated by statsBordeaux package
}
\examples{
data("mtcars")
mtcars$vs <- as.factor(mtcars$vs)
table <- statsBordeaux::setMultivariableRegression(data = mtcars,
                                                   dependentVariable = "mpg",
                                                   independentVariables = c("cyl", "disp", "vs"),
                                                   round = 2)
doc <- read_docx()
doc <- addTable(doc, table, "Description of mpg variable from the mtcars dataset")
print(doc, target = "output.docx")
}
