% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{addTable}
\alias{addTable}
\title{Add table into a rdocx element}
\usage{
addTable(
  doc,
  table,
  title,
  displayTestName = TRUE,
  modalitySize = 1,
  descriptionSize = 1.1,
  valueSize = 1,
  testSize = 2
)
}
\arguments{
\item{doc}{a rdocx element}

\item{table}{a data.frame generated by statsBordeaux package}

\item{title}{a character vector of length one, the title of the table to display}

\item{displayTestName}{boolean, if TRUE, name of the statistic test will be displayed. Default to TRUE}

\item{modalitySize}{double, size in inch of the modality column}

\item{descriptionSize}{double, size in inch of the description column}

\item{valueSize}{double, size in inch of the value column}

\item{testSize}{double, size in inch of the test name column}
}
\value{
a rdocx element with the table at the end of the document
}
\description{
Add table generated by statsBordeaux package
}
\examples{
data(mtcars)
output <- createOutput()
output <- statsQT(output, mtcars, "mpg")
doc <- read_docx()
doc <- addTable(doc, output, "Description of mpg variable from the mtcars dataset")
print(doc, target = "output.docx")
}
