% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{addTableOfContent}
\alias{addTableOfContent}
\title{Add table of content to rdocx element}
\usage{
addTableOfContent(doc, title, style, level)
}
\arguments{
\item{doc}{a rodcx element}

\item{title}{a character vector of length one, the title of the table of content}

\item{style}{a character vector of length one, the style of the title of the table of content. Styles could
be found with the styles_info(doc) function}

\item{level}{a numeric vector of length one, the deep of the tithe hierarchy to display}
}
\value{
a rdocx element with the table of content
}
\description{
Add table of content to rdocx element
}
\examples{
doc <- read_docx()
doc <- addTableOfContent(doc, title = "Table of content", style = "Normal", level = 3)
doc <- addBreak(doc)
doc <- addTextElement(doc, "Titre 1 : description of the dataset", style = "heading 1")
doc <- addTextElement(doc, "Titre 2 : comparaison of the dataset", style = "heading 1")
doc <- addTextElement(doc, "Titre 2.1 : comparaison of the dataset depend on the sexe variable", style = "heading 2")
doc <- addTextElement(doc, "Titre 2.2 : description of the dataset depend on the group variable", style = "heading 2")
print(doc, target = "output.docx")
}
