% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{addTableOfGraphics}
\alias{addTableOfGraphics}
\title{Add a table of graphics into a rdocx element}
\usage{
addTableOfGraphics(doc, title, style)
}
\arguments{
\item{doc}{a rdocx element}

\item{title}{a character vector of length one, the title of the table of graphics}

\item{style}{a character vector of length one, the style of the title of the table of graphics}
}
\value{
a rdocx element with a table of graphics
}
\description{
Add a table of graphics into a rdocx element
}
\examples{
doc <- read_docx()
doc <- addTextElement(doc, "Titre 1 : description of the dataset", style = "heading 1")
doc <- addNewLine(doc)
doc <- addNewLine(doc)
doc <- addPlot(doc, plot, "mpg function of vs")
doc <- setLandscape(doc)
doc <- addTableOfGraphics(doc, title = "Table of graphics", style = "Normal")
print(doc, target = "output.docx")
}
