% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{addTableOfTables}
\alias{addTableOfTables}
\title{Add a table of tables into a rdocx element}
\usage{
addTableOfTables(doc, title, style)
}
\arguments{
\item{doc}{a rdocx element}

\item{title}{a character vector of length one, the title of the table of tables}

\item{style}{a character vector of length one, the style of the title of the table of tables}
}
\value{
a rdocx element with a table of table
}
\description{
Add a table of tables into a rdocx element
}
\examples{
data(mtcars)
output <- createOutput()
output <- statsQT(output, mtcars, "mpg")
doc <- read_docx()
doc <- addTable(doc, output, "Description of mpg variable from the mtcars dataset")
doc <- addBreak(doc)
doc <- addTableOfTables(doc, title = "Table of tables", style = "Normal")
print(doc, target = "output.docx")
}
