% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{addTextElement}
\alias{addTextElement}
\title{Add text element into the document}
\usage{
addTextElement(doc, title, style)
}
\arguments{
\item{doc}{a rdocx element}

\item{title}{a character vector of length 1, the text elemen to display}

\item{style}{a character vector of length 1. Style coult be get executing the function styles_info(doc),
in the style_name column}
}
\value{
an rdocx element with the text element at the end
}
\description{
Add text element into the document

Add text element into the document
}
\examples{
doc <- read_docx()
doc <- addTextElement(doc, "Titre 1 : description of the dataset", style = "heading 1")
print(doc, target = "output.docx")
}
