% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{checkNormality}
\alias{checkNormality}
\title{Test normality of numeric variable based on graphic and statistc methods}
\usage{
checkNormality(
  data,
  variable = colnames(data),
  group = NULL,
  p_value = FALSE,
  method = "Kolmogorov"
)
}
\arguments{
\item{data}{a data.frace conteting the numeric data witch normality need to be evaluate}

\item{variable}{a character vector. The name of the numeric columns to describe.}

\item{group}{a character vector of length 1. The name of the factor column to use as commparaison group. Default to NULL.}

\item{p_value}{a boolean. If TRUE, comparaison test to normality are performed. Default to FALSE.}

\item{method}{a character vector of length 1. The name og the statistic method to use : 'Kolmogorov' or 'Shapiro'}
}
\value{
a plot
}
\description{
Test normality of numeric variable based on graphic and statistc methods
}
\examples{
data(mtcars)
checkNormality(mtcars, "wt", p_value = TRUE)
}
