% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{checkNotDigitInDataframe}
\alias{checkNotDigitInDataframe}
\title{Check if data are only digits}
\usage{
checkNotDigitInDataframe(data, returnError = FALSE, excelFormat = TRUE)
}
\arguments{
\item{data}{a data.frame containing the data to check}

\item{returnError}{boolean. If TRUE, return a data.frame with the coordinate of the cell}

\item{excelFormat}{a boolean vector of length one. If TRUE, data.frame containing position of none
digit cell is in excel format. Default to TRUE.}
}
\value{
TRUE if the data.frame only contain digit, else return a data.frame
with the coordinate of the cell containing some wrond format data in excel style. FALSE by default.
}
\description{
Check if the input data.frale are only composed of digits
}
\examples{
data(mtcars)
checkNotDigitInDataframe(data = mtcars, returnError = FALSE)
mtcars$NOT_DIGIT <- rep("A", nrow(mtcars))
errorPosition <- checkNotDigitInDataframe(data = mtcars, returnError = TRUE)
}
