% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{createOutput}
\alias{createOutput}
\title{create a data.frame with col and row as specified}
\usage{
createOutput(ncol = 0, nrow = 0)
}
\arguments{
\item{ncol}{a digit. Default to 0}

\item{nrow}{a digit. Default to 0}
}
\value{
a data.frame
}
\description{
create a data.frame withe number of columns and rows as specicied
}
\examples{
result <- createOutput(ncol = ncol(mtcars))
colnames(result) <- colnames(mtcars)
}
