% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{describeDataFrame}
\alias{describeDataFrame}
\title{A convenient method to describe a full data.frame.}
\usage{
describeDataFrame(
  data,
  variables = colnames(data),
  group = NULL,
  group_str = NULL,
  p_value = FALSE,
  all = FALSE,
  desc = c("Mean", "Median", "Range"),
  round = 3,
  confint = FALSE,
  NA_asModality = FALSE,
  NA_group_AsModality = FALSE
)
}
\arguments{
\item{data}{a data.frame containing the data to describe}

\item{variables}{a character vector contening the name of columns to describe. Default to colnames(data).}

\item{group}{a character vector of length 1. The name of the factor column to use as commparaison group. Default to NULL.}

\item{group_str}{a numeric vector. The index of the levels of the group variable to use. Default to NULL.}

\item{p_value}{a boolean. If TRUE, comparaison test are performed.}

\item{all}{a boolean. If TRUE, total column will be displayed. Default to FALSE}

\item{desc}{a character vector. Could contain "Mean", "Median", "Range" and/or "Mode"}

\item{round}{an integer, number of maximal decimal. Default to 3}

\item{confint}{a boolean. If TRUE, the confidence interval of the mean will be displayed. Default to FALSE}

\item{NA_asModality}{a boolean. If TRUE, missing data of the factor variable to describe will be considered as levels.
Default to FALSE}

\item{NA_group_AsModality}{a boolean. If TRUE, missing data of the group variable will be considered as levels. Default to FALSE}
}
\value{
a data.frame containing the description of the variables
}
\description{
A convenient method to describe a full data.frame.
}
\examples{
data(mtcars)
labels <- data.frame(Variable = c("vs", "vs", "am", "am"),
                     Modality = c(0, 1, 0, 1),
                     Label = c("V-shaped", "Straight", "Automatic", "Manual"))
labelVariable <- data.frame(Variable = c("mpg", "cyl", "disp", "hp", "drat", "wt",
                                         "qsec", "vs", "am", "gear", "carb"),
                            Label = c("Miles/(US) gallon", "Number of cylinders", "Displacement (cu.in.)",
                                      "Gross horsepower ", "Rear axle ratio", "Weight (1000 lbs)",
                                      "1/4 mile time", "Engine", "Transmission", "Number of forward gears",
                                      "Number of carburetors"))
labelledData <- statsBordeaux::labellisationDataFrame(mtcars, labels)
labelledData <- statsBordeaux::setLabelToVariable(labelledData, labelVariable)
comparaison <- describeDataFrame(mtcars, group = "vs", p_value = TRUE)
}
