% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_function.R
\name{diag.kappa.getAgreeentMatrix}
\alias{diag.kappa.getAgreeentMatrix}
\title{Display agreement matrix table in some Rmarkdown document}
\usage{
diag.kappa.getAgreeentMatrix(diag.kappa)
}
\arguments{
\item{diag.kappa}{a list, result of diag.kappa() function}
}
\value{
a kableExtra table
}
\description{
Display agreement matrix table produces by diag.kappa() in some Rmarkdown document
}
