% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_function.R
\name{diag.kappa}
\alias{diag.kappa}
\title{Perform some kappa test between two qualitative test}
\usage{
diag.kappa(data, test_1, test_2, round = 3)
}
\arguments{
\item{data}{a data.frame, containing the tests' result as factor}

\item{test_1}{a character vector of length one, containing the colname
of the test_1 procedure}

\item{test_2}{a character vector of length one, containing the colname
of the test_2 procedure}

\item{round}{an integer, number of maximal decimal. Default to 3}
}
\value{
a list contaning agreement matrix table and agreement parameters as data.frame
}
\description{
Perform some kappa test between two qualitative test
}
