% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_function.R
\name{diag.perf.getContingencyTable}
\alias{diag.perf.getContingencyTable}
\title{Display contengency table in some Rmarkdown document}
\usage{
diag.perf.getContingencyTable(diag.perf)
}
\arguments{
\item{diag.perf}{a list, result of diag.perf() function}
}
\value{
a kableExtra table
}
\description{
Display contengency table produce with diag.perf() function in some Rmarkdown document
}
