% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_function.R
\name{diag.perf.getPerformanceTable}
\alias{diag.perf.getPerformanceTable}
\title{Display performance table in some Rmarkdown document}
\usage{
diag.perf.getPerformanceTable(diag.perf)
}
\arguments{
\item{diag.perf}{a list, result of diag.perf() function}
}
\value{
a kableExtra table
}
\description{
Display performance table produce with diag.perf() function in some Rmarkdown document
}
