% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_function.R
\name{diag.perf}
\alias{diag.perf}
\title{Get diagnostic performance evaluation of some qualitative procedure}
\usage{
diag.perf(data, goldStandard, indexTest, M_pos, T_pos, round = 3)
}
\arguments{
\item{data}{a data.frame, containing the diagnostic result}

\item{goldStandard}{a character vector of length one, containing the colname
of the gold-standard procedure}

\item{indexTest}{a character vector of length one, containing the colname
of the index procedure}

\item{M_pos}{a character vector, containing the levels of positive result
with gold-standard procedure}

\item{T_pos}{a character vector, containing the levels of positive result
with index procedure}

\item{round}{an integer, number of maximal decimal. Default to 3}
}
\value{
a list contaning contengency table and performance evaluation as data.frame
}
\description{
Get diagnostic performance evaluation of some qualitative test
}
