% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{firstUp}
\alias{firstUp}
\title{Capitalize the first letter of a string}
\usage{
firstUp(x)
}
\arguments{
\item{x}{a character vector of length one}
}
\value{
a character vector of length one, with the first letter capitalize
}
\description{
Capitalize the first letter of a string
}
