% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{getBarPlot}
\alias{getBarPlot}
\title{A convenient method to get bar plot from qualitative variable.}
\usage{
getBarPlot(data, variable, group = NULL, legend.position = "right",
  na.rm = TRUE)
}
\arguments{
\item{data}{a data.frame, containing the qualitative variable to describe with barplot}

\item{variable}{a character vector of length one, containing the name of the
variable to describe}

\item{group}{a character vector of length one, containing the name of the
group which need to be used to describe the qualitative variable. Depault to NULL}

\item{legend.position}{a characher vector of length one, containing the legend position.
Must be in 'right', 'left', 'top', 'bottom, 'none'. Default to 'right'.}

\item{na.rm}{a boolean vector of length one. If TRUE, missig value of variable will be remove.
Default to TRUE}
}
\value{
a ggplot2 plot
}
\description{
A convenient method to get bar plot from qualitative variable.
}
\examples{
data(mtcars)
mtcars$am <- as.factor(mtcars$am)
mtcars$vs <- as.factor(mtcars$vs)
getBarPlot(data = mtcars, variable = "vs", group = "am")
}
