% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{getBoxPlot}
\alias{getBoxPlot}
\title{A convenient method to get box plot from numeric variable.}
\usage{
getBoxPlot(data, variable, group = NULL, legend.position = "right")
}
\arguments{
\item{data}{a data.frame, containing the numeric variable to describe with boxplot}

\item{variable}{a character vector, containing the name of the variables to describe}

\item{group}{a character vector of length one, containing the name of the
group which need to be used to describe the numeric variable. Depault to NULL}

\item{legend.position}{a characher vector of length one, containing the legend position.
Must be in 'right', 'left', 'top', 'bottom, 'none'. Default to 'right'.}
}
\value{
a ggplot2 plot
}
\description{
A convenient method to get box plot from numeric variable.
}
\examples{
data(mtcars)
mtcars$am <- as.factor(mtcars$am)
getBoxPlot(data = mtcars, variable = "mpg", group = "am")
}
