% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{getCorrelationCoeficient}
\alias{getCorrelationCoeficient}
\title{Get correlation coeficient between two numeric variable}
\usage{
getCorrelationCoeficient(
  data,
  var1,
  var2,
  method = "pearson",
  exact = NULL,
  conf.level = 0.95,
  round = 3
)
}
\arguments{
\item{data}{a data.frame, containing the data to use to compute the correlation coeficient}

\item{var1}{a character vector of length one, containing the variable name of numeric variable in data}

\item{var2}{a character vector of length one, containing the variable name of numeric variable in data}

\item{method}{a character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman".}

\item{exact}{a logical indicating whether an exact p-value should be computed. Used for Kendall's tau and Spearman's rho.}

\item{conf.level}{confidence level for the returned confidence interval.}

\item{round}{a numeric vector of length one, containing the number of digit for rounding}
}
\value{
a data.frame
}
\description{
Get correlation coeficient between two numeric variable
}
\examples{
data(mtcars)
cor <- getCorrelationCoeficient(data = mtcars, var1 = 'mpg', var2 = colnames(mtcars)[sapply(mtcars, is.numeric)])
cor <- do.call('rbind', cor)
}
