% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{getDensityPlot}
\alias{getDensityPlot}
\title{Get density plot of numeric variable into data.frame}
\usage{
getDensityPlot(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
a list of plot
}
\description{
Get density plot of numeric variable into data.frame
}
\examples{
data(mtcars)
getDensityPlot(mtcars)
}
