% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{getGraphicalDescription}
\alias{getGraphicalDescription}
\title{A convenient method to get plot from data.frame.}
\usage{
getGraphicalDescription(data, variable = colnames(data), group = NULL,
  legend.position = "right", na.rm = TRUE)
}
\arguments{
\item{data}{a data.frame, containing the numeric variable to describe with boxplot}

\item{variable}{a character vector, containing the name of the variables to describe}

\item{group}{a character vector of length one, containing the name of the}

\item{legend.position}{a characher vector of length one, containing the legend position.
Must be in 'right', 'left', 'top', 'bottom, 'none'. Default to 'right'.}

\item{na.rm}{a boolean vector of length one. If TRUE, missig value of variable will be remove.
Default to TRUE}
}
\value{
a ggplot2 plot
}
\description{
A convenient method to get plot from data.frame
}
\examples{
data(mtcars)
mtcars$am <- as.factor(mtcars$am)
mtcars$vs <- as.factor(mtcars$vs)
mtcars \%>\% getGraphicalDescription(group = "am")
}
