% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{getVarLabel}
\alias{getVarLabel}
\title{Get the label of variable in a data.frame}
\usage{
getVarLabel(data, variable = colnames(data))
}
\arguments{
\item{data}{a data.frame containing colname to get variable label}

\item{variable}{a character vector, containing the colnames to querry}
}
\value{
a named vector of the label of each variable
}
\description{
Get the label of variables witch were labelised thanks the setLabel function
}
\examples{
data(mtcars)
label <- data.frame(VARIABLE = colnames(mtcars),
                    LABEL = c("Miles/(US) gallon",
                              "Number of cylinders",
                              "Displacement (cu.in.)",
                              "Gross horsepower",
                              "Rear axle ratio",
                              "Weight (1000 lbs)",
                              "1/4 mile time",
                              "Engine (0 = V-shaped, 1 = straight)",
                              "Transmission (0 = automatic, 1 = manual)",
                              "Number of forward gears",
                              "Number of carburetors"),
                    stringsAsFactors = FALSE)

labeled_df <- setLabelToVariable(mtcars, label)
getVarLabel(labeled_df)
}
