% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isCastableAsNumeric}
\alias{isCastableAsNumeric}
\title{A conveinineg metho to check if some vector could be cast as numeric}
\usage{
isCastableAsNumeric(vector)
}
\arguments{
\item{vector}{a vector, which need to be cast as numeric}
}
\value{
a boolean vector of length 1.
}
\description{
A conveinineg metho to check if some vector could be cast as numeric
}
