% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{labellisationDataFrame}
\alias{labellisationDataFrame}
\title{set label to qualitative variable into a data.frame based on dictionnary.}
\usage{
labellisationDataFrame(df, labelTable)
}
\arguments{
\item{df}{a data.frame containing qualitative variable witch need to be labelised.}

\item{labelTable}{a data.frame containing three columns : 1) the variable names, 2) one
modality as digit, 3) the label of the modality.}
}
\value{
a data.frame with qualitative variable in the labelTable convert as factor with clean label
}
\description{
manage the conversion and sat label to qualitative into a data.frame based
on dictionnary.
}
\examples{
df <- data.frame(X1 = c(0, 1, 2, 1, 2, 1, 2),
X2 = c(0, 1, 0, 1, 1, 0, 0))
labelTable <- data.frame(VARIABLE = c("X2", "X2"),
                         MODALITY = c(0, 1),
                         LABEL = c("Male", "Female"))
df <- labellisationDataFrame(df, labelTable)
}
