% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.R
\name{lang.available}
\alias{lang.available}
\title{Get all available languages}
\usage{
lang.available()
}
\value{
a character vector containing all available languages
}
\description{
Get all available languages
}
