% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{manageMultipleColumnVariableAndDictionnary}
\alias{manageMultipleColumnVariableAndDictionnary}
\title{A conveinient method to manage multiple column data and label dictionnary}
\usage{
manageMultipleColumnVariableAndDictionnary(data, labelVar, labelQL,
  listVarToMerge, nonAppChar = "NonApp", varOutputLabel)
}
\arguments{
\item{data}{a data.frame, containing the data with the indicator variable created by manageMultipleColumnVariable}

\item{labelVar}{a data.frame with two columns : 1) the variable colname, 2) the clean label to display}

\item{labelQL}{a data.frame containing three columns : 1) the variable names, 2) one modality as digit, 3) the label of the modality.}

\item{listVarToMerge}{a character vector, containing the colname of the column to merge}

\item{nonAppChar}{a character vector of length one, containing the colname of the column to merge}

\item{varOutputLabel}{a character vector of length one, containing the prefix of the columns to create}
}
\value{
a list of three data.frame : 1) data, 2) labelVar, 3) labelQL
}
\description{
A conveinient method to manage multiple column data and label dictionnary
}
