% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{manageNotApplicable}
\alias{manageNotApplicable}
\title{Manage the not applicable condition into data.frame}
\usage{
manageNotApplicable(data, notApplicableChar)
}
\arguments{
\item{data}{a data.frame containing data to clean}

\item{notApplicableChar}{a character vector of length 1, with the not applicable character as in the data.frame}
}
\value{
a list of two elements : 1) data.frame without not applicable character, 2) list of logical vector
identifying the row witch should contain data
}
\description{
Identify the not applicable char in data.frame and return clean list with :
1) data.frame without not applicable character, 2) list of logical vector identifying the row witch should contain data
}
\examples{
data <- data.frame(X1 = c(1, 1, 0, 1, 1, 0, 1, 1, 1),
                   X2 = c(0, 0, "NC", 1, 1, "NC", 0, 0, 1))
resultNotApplicable <- manageNotApplicable(data, "NC")
data <- resultNotApplicable[[1]]
resultNotApplicable <- resultNotApplicable[[2]]
}
