% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{mergePairedDataFrame}
\alias{mergePairedDataFrame}
\title{A convenient method that merge two data.frame of paired data}
\usage{
mergePairedDataFrame(dataA, dataB, byA, byB, nameA, nameB)
}
\arguments{
\item{dataA}{a data.frame, containing some data to analyse}

\item{dataB}{a data.frame, containing some data to analyse}

\item{byA}{a character vector of length one, containing the name of the column in dataA used in the merge}

\item{byB}{a character vector of length one, containing the name of the column in dataB used in the merge}

\item{nameA}{a character vector of length one, containing the name of the group of dataA}

\item{nameB}{a character vector of length one, containing the name of the group of dataB}
}
\value{
a data.frame containing the paired data to analyze
}
\description{
A convenient method than merge two data.frame of paired data and return a data.frame
in the good format to be analyzed
}
