% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restoreAttributes}
\alias{restoreAttributes}
\title{Save column's attribute to a data.frame}
\usage{
restoreAttributes(data = ., listAttributes = c("label", "var_label"))
}
\arguments{
\item{data}{a data.frame}

\item{listAttributes}{the list of attribute to restore}
}
\value{
a data.frame with the column's attribute restored
}
\description{
Save column's attribute to a data.frame
}
