% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveAttributes}
\alias{saveAttributes}
\title{Save column's attribute to a data.frame}
\usage{
saveAttributes(data = .)
}
\arguments{
\item{data}{a data.frame}
}
\value{
a data.frame with the column's attribute saved
}
\description{
Save column's attribute to a data.frame
}
