% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{setFormatToNumber}
\alias{setFormatToNumber}
\title{Format some digit in the french format for display usage}
\usage{
setFormatToNumber(num, round = 3)
}
\arguments{
\item{round}{an integer, the number of digit after the decimal separator. Default to 3}

\item{pvalues}{a numeric, witch need to be formated}
}
\value{
a formated character
}
\description{
Manage the display digit format in french format
}
\examples{
digit <- 12.359
digit_f <- setFormatToNumber(digit)
digit_f
}
