% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{setFormatToPvalue}
\alias{setFormatToPvalue}
\title{Format some float in the french p-value format for display usage}
\usage{
setFormatToPvalue(pvalue, round = 3)
}
\arguments{
\item{round}{an integer, the number of digit after the decimal separator. Default to 3}

\item{pvalues}{a numeric, witch need to be formated}
}
\value{
a formated character
}
\description{
Manage the display of p-value format in french format
}
\examples{
pvalue <- 0.00028
pvalue_f <- setFormatToPvalue(pvalue)
pvalue_f
}
