% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{setLabelToFactorLevels}
\alias{setLabelToFactorLevels}
\title{set label to qualitative variable into a data.frame based on dictionnary.}
\usage{
setLabelToFactorLevels(
  data,
  factorLevelsLabel,
  varIndex = 1,
  levelsIndex = 2,
  labelIndex = 3,
  levelsThreesholdCheck = 5
)
}
\arguments{
\item{data}{a data.frame containing qualitative variable witch need to be labelised.}

\item{factorLevelsLabel}{a data.frame containing variable names / modality as digit / label of the modality.}

\item{varIndex}{a numeric vector of length one, containing the index of the variable column.}

\item{levelsIndex}{a numeric vector of length one, dontaining the index of the levels column.}

\item{labelIndex}{a numeric vector of length one, containing the index of the labels column.}

\item{levelsThreesholdCheck}{a numeric vector of length one, containing the disctinct
number value threeshold to check to evaluate if all factor variable are listed in the
factorLevelsLabel table}
}
\value{
a data.frame with qualitative variable in the factorLevelsLabel convert as factor with clean label
}
\description{
manage the conversion and sat label to qualitative into a data.frame based
on dictionnary.
}
\examples{
data <- data.frame(X1 = c(0, 1, 2, 1, 2, 1, 2),
X2 = c(0, 1, 0, 1, 1, 0, 0))
labelTable <- data.frame(VARIABLE = c("X2", "X2"),
                         MODALITY = c(0, 1),
                         LABEL = c("Male", "Female"))
data <- setLabelToFactorLevels(data, labelTable)
}
