% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{setLabelToVariable}
\alias{setLabelToVariable}
\title{set label to each column of data.frame based on dictionnaty}
\usage{
setLabelToVariable(data, varLabel, varIndex = 1, labelIndex = 2)
}
\arguments{
\item{data}{the data.frame witch nedd to be labelised}

\item{varLabel}{a data.frame with two columns : 1) the variable colname, 2) the clean label to display}

\item{varIndex}{a numeric vector of length one, containing the index if the variable label column.}

\item{labelIndex}{a numeric vector of length one, containing the index if the variable label column.}
}
\value{
a data.frame with the var_label attribute
}
\description{
add a label as attribute (var_label) to each column of data.frame based on dictionnary. This
label will be used during the reporting function.
}
\examples{
data(mtcars)
label <- data.frame(VARIABLE = colnames(mtcars),
                    LABEL = c("Miles/(US) gallon",
                              "Number of cylinders",
                              "Displacement (cu.in.)",
                              "Gross horsepower",
                              "Rear axle ratio",
                              "Weight (1000 lbs)",
                              "1/4 mile time",
                              "Engine (0 = V-shaped, 1 = straight)",
                              "Transmission (0 = automatic, 1 = manual)",
                              "Number of forward gears",
                              "Number of carburetors"),
                    stringsAsFactors = FALSE)

labeled_df <- setLabelToVariable(mtcars, label)
lapply(labeled_df, function(x) attributes(x)$`var_label`)
}
