% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{setLandscape}
\alias{setLandscape}
\title{Set to landscape the previsous section}
\usage{
setLandscape(doc, add_break = FALSE)
}
\arguments{
\item{doc}{a rdocx element}

\item{add_break}{a boolean, if TRUE, add break after the landscape section.}
}
\value{
a rdocx element with the last section as landscape
}
\description{
Set to previsous the previsous section
}
\examples{
doc <- read_docx()
doc <- addTextElement(doc, "Titre 1 : description of the dataset", style = "heading 1")
doc <- addNewLine(doc)
doc <- addNewLine(doc)
doc <- addPlot(doc, plot, "mpg function of vs")
doc <- setLandscape(doc)
print(doc, target = "output.docx")
}
