% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_function.R
\name{setMultivariableRegression}
\alias{setMultivariableRegression}
\title{Perform univariable regression models}
\usage{
setMultivariableRegression(data, dependentVariable, independentVariables,
  round = 2)
}
\arguments{
\item{data}{a data.frame, containing the data to be used in the regression}

\item{dependentVariable}{a character vector of length one, containing the names of the dependant variable}

\item{independentVariables}{a character vector containing the names of the independant variables}

\item{round}{a numeric vector of length one, containing the round value to apply}
}
\value{
a data.frame, containing the result of the regression model
}
\description{
A fonction that allow to perform multivariable linear and logistic regression and
produce a data.frame to display
}
\examples{
data("mtcars")
setMultivariableRegression(data = mtcars,
                           dependentVariable = "cyl",
                           independentVariables = c("mpg", "drat")
}
