% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setNonApp}
\alias{setNonApp}
\title{Set non-applicable tag depend on condition}
\usage{
setNonApp(data = ., colname, condition, notApplicableChar = "NonApp")
}
\arguments{
\item{data}{a data.frame}

\item{colname}{a character vector of length one, in which non-applicable tag must be set}

\item{condition}{the condition that trigger the set of non-applicable tag}

\item{notApplicableChar}{a character vector of length one, containing the non-applicable tag. Default to 'NonApp'}
}
\value{
a data.frame with the non-applicable tag set
}
\description{
A convenient method to set non-applicable tag depend on condition
}
