% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_function.R
\name{setUnivariableRegression}
\alias{setUnivariableRegression}
\title{Perform univariable regression models}
\usage{
setUnivariableRegression(
  data,
  dependentVariable,
  independentVariable,
  round = 2
)
}
\arguments{
\item{data}{a data.frame, containing the data to be used in the regression}

\item{dependentVariable}{a character vector of length one, containing the names of the dependant variable}

\item{independentVariable}{a character vector of length one, containing the names of the independant variable}

\item{round}{a numeric vector of length one, containing the round value to apply}
}
\value{
a data.frame, containing the result of the regression model
}
\description{
A fonction that allow to perform univariable linear and logistic regression and
produce a data.frame to display
}
\examples{
data("mtcars")
setUnivariableRegression(data = mtcars,
                         dependentVariable = "cyl",
                         independentVariable = "mpg")
}
