% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsBordeaux-package.R
\docType{package}
\name{statsBordeaux-package}
\alias{statsBordeaux}
\alias{statsBordeaux-package}
\title{statsBordeaux: Helpful tool to describe and analyse dataset}
\description{
Method to describe and analyse dataset and generate reporting.
}
\keyword{internal}
