% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{statsQL}
\alias{statsQL}
\title{Generate description of factor varaible.}
\usage{
statsQL(data, variable, group = NULL, group_str = NULL,
  all = NA_asModality, round = 3, p_value = FALSE,
  forcedTest = NULL, NA_group_AsModality = FALSE,
  NA_asModality = FALSE, output = NULL)
}
\arguments{
\item{data}{a data.frame containing the data to describe}

\item{variable}{a character vector of length 1. The name of the factor column to describe.}

\item{group}{a character vector of length 1. The name of the factor column to use as commparaison group. Default to NULL.}

\item{group_str}{a character vector. The name of the levels of the group variable to use. Default to NULL.}

\item{all}{a boolean. If TRUE, total column will be displayed. Default to FALSE}

\item{round}{an integer, number of maximal decimal. Default to 3}

\item{p_value}{a boolean. If TRUE, comparaison test are performed.}

\item{forcedTest}{a character vector of length 1. Must be one of "Pearson's Chi-squared test", "Pearson's Chi-squared test with Yates" or
"Fisher's Exact Test for Count Data". Default to NULL.}

\item{NA_group_AsModality}{a boolean. If TRUE, missing data of the group variable will be considered as levels.
Default to FALSE}

\item{NA_asModality}{a boolean. If TRUE, missing data of the factor variable to describe will be considered as levels.
Default to FALSE}

\item{output}{a data.frame, containing a previously generated data.frame or a new data.frame.
data.frame are generated using createOutput() function.}
}
\value{
a data.frame containing the description of the factor variable
}
\description{
A method that generate description of factor variable and perform comparative statistic test
in case of comparaison group.
}
\examples{
data(mtcars)
mtcars$am <- as.factor(mtcars$am)
mtcars$vs <- as.factor(mtcars$vs)
output <- createOutput()
output <- statsQT(output = output,
                  input = mtcars,
                  variable = "mpg",
                  group = "am")
output <- statsQL(output = output,
                  input = mtcars,
                  variable = "vs",
                  group = "am")
}
