% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{statsQT}
\alias{statsQT}
\title{Generate description of numeric varaible.}
\usage{
statsQT(output, input, variable, group = NULL, group_str = NULL,
  p_value = FALSE, forcedTest = NULL, all = FALSE, desc = c("Mean",
  "Median", "Range"), round = 3, confint = FALSE,
  NA_group_AsModality = FALSE)
}
\arguments{
\item{output}{a data.frame, containing a previously generated data.frame or a new data.frame.
data.frame are generated using createOutput() function.}

\item{input}{a data.frame containing the data to describe}

\item{variable}{a character vector of length 1. The name of the numeric column to describe.}

\item{group}{a character vector of length 1. The name of the factor column to use as commparaison group. Default to NULL.}

\item{group_str}{a character vector. The name of the levels of the group variable to use. Default to NULL.}

\item{p_value}{a boolean. If TRUE, comparaison test are performed.}

\item{forcedTest}{a character vector of length 1. Must be one of "Student", "Welch", "Wilcoxon", "ANOVA", or
"Kruskal-Wallis". Default to NULL.}

\item{all}{a boolean. If TRUE, total column will be displayed. Default to FALSE}

\item{desc}{a character vector. Could contain "Mean", "Median", "Range" and/or "Mode"}

\item{round}{an integer, number of maximal decimal. Default to 3}

\item{confint}{a boolean. If TRUE, the confidence interval of the mean will be displayed. Default to FALSE}

\item{NA_group_AsModality}{a boolean. If TRUE, missing data of the group variable will be considered as levels.
Default to FALSE}
}
\value{
a data.frame containing the description of the variable
}
\description{
A method that generate description of numeric variable and perform comparative statistic test
in case of comparaison group.
}
\examples{
data(mtcars)
mtcars$am <- as.factor(mtcars$am)
output <- createOutput()
output <- statsQT(output = output,
                  input = mtcars,
                  variable = "mpg",
                  group = "am")
output <- statsQT(output = output,
                  input = mtcars,
                  variable = "disp",
                  group = "am")
}
