% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{subset_withAttributes}
\alias{subset_withAttributes}
\title{Subset data keeping attributes}
\usage{
subset_withAttributes(data, subset)
}
\arguments{
\item{data}{a data.frame}

\item{subset}{a logical vector}
}
\value{
a data.frame with all the attributes
}
\description{
Allows to subset data and keeping all the attributes of the data
}
\examples{
data(mtcars)
mtcars$nSpeed <- c(5, 6, 5, "NA", "NA", "NA", "NA", "NA", "NA", "NA",
                  "NA", "NA", "NA", "NA", "NA", "NA", "NA", 6, 6, 6,
                  "NA", "NA", "NA", "NA", "NA", 5, 5, 6, 5, 6, 6, 5)
attributes(mtcars[, "nSpeed"])$var_label <- "Number of speed in case of manual transmission"
subset <- subset_withAttributes(mtcars, mtcars$nSpeed != "NA")
attributes(subset[, "nSpeed"])
}
