% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update.R
\name{updateStatsBordeaux}
\alias{updateStatsBordeaux}
\title{Update the statsBordeaux package}
\usage{
updateStatsBordeaux(force = FALSE)
}
\arguments{
\item{force}{a boolean vector of length one. If TRUE, force installation, even if the
remote state has not changed since the previous install. Default to FALSE}
}
\description{
Update the statsBordeaux package
}
\examples{
updateStatsBordeaux()
}
