% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_function.R
\name{addKable}
\alias{addKable}
\title{A convenient method to add kable into a RMarkdown document}
\usage{
addKable(data_frame, all_before = FALSE)
}
\arguments{
\item{data_frame}{a data.frame, containing the data to add}
}
\value{
a knitr::kable
}
\description{
A convenient method to add kable into a RMarkdown document
}
\examples{
data(mtcars)
output <- createOutput()
output <- statsQT(output, mtcars, "mpg")
addKable(output)
}
