% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{checkNotDigitInDataframe}
\alias{checkNotDigitInDataframe}
\title{Check if data are only digits}
\usage{
checkNotDigitInDataframe(df, returnError = FALSE)
}
\arguments{
\item{df}{the data.frame to check}

\item{returnError}{boolean. If TRUE, return a data.frame with the coordinate of the cell
containing some non digit data in excel style.}
}
\value{
TRUE if the data.frame only contain digit, else return a data.frame
with the coordinate of the cell containing some wrond format data in excel style. FALSE by default.
}
\description{
Check if the input data.frale are only composed of digits
}
\examples{
data(mtcars)
checkNotDigitInDataframe(df = mtcars, returnError = FALSE)

mtcars$NOT_DIGIT <- rep("A", nrow(mtcars))
errorPosition <- checkNotDigitInDataframe(df = mtcars, returnError = TRUE)
}
